start triggers domain layer implementation: create Trigger.kt data class create TriggersRepository.kt interface
update WeatherModule.kt:
add provides method for TriggersApi

refactor TriggerResponseDto.kt and TriggerRequestDto.kt:
change condition amount type from Int to Double

+++ app/src/main/java/com/example/weatherapp/data/remote/weather/dto/TriggerRequestDto.kt
@@ -21,7 +21,7 @@ data class TriggerRequestTimeExpression(
 data class TriggerRequestCondition(
     @SerializedName("name") val name: String,
     @SerializedName("expression") val expression: String,
-    @SerializedName("amount") val amount: Int,
+    @SerializedName("amount") val amount: Double,
 )
 
 data class TriggerRequestArea(
